requirements.txt
PyMySQL~=1.1.1
aiomysql~=0.2.0
mashumaro[yaml]~=3.15
python-telegram-bot[job-queue]~=21.11.1
